perf(寄存器组): 调整模块结构
